Data set buildet now builds into actual training set and starts neural network thingies. Shit's real now

+++ Master Offline Tool/Master Offline Tool/DataSetBuilder.h
@@ -11,7 +11,7 @@ struct DataSet
 	int index;
 	int inputs;
 	std::vector<float> values;
-	int output;
+	float output;
 };
 
 class DataSetBuilder
